+++ lib/src/app.dart
@@ -5,7 +5,6 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../main.dart';
 import 'constants.dart';
-// import 'screens/history.dart';
 import 'qr_image_config.dart';
 import 'screens/home.dart';
