+++ lib/src/constants.dart
@@ -7,16 +7,12 @@ enum DefaultPopupMenuItemsType {
   history,
   about,
   selectTheme,
-  // privacyPolicy,
-  // terms
 }
 
 const List<Map<String, dynamic>> defaultPopupMenuItems = [
   {'label': '履歴', 'value': DefaultPopupMenuItemsType.history},
   {'label': 'このアプリについて', 'value': DefaultPopupMenuItemsType.about},
   {'label': 'テーマの選択', 'value': DefaultPopupMenuItemsType.selectTheme},
-  // {'label': 'プライバシーポリシー', 'value': DefaultPopupMenuItemsType.privacyPolicy},
-  // {'label': '利用規約', 'value': DefaultPopupMenuItemsType.terms}
 ];
 
 const defaultSheetHeight = 0.2;
@@ -29,47 +25,6 @@ class SelectOption<T> {
   SelectOption({required this.label, required this.value, this.description});
 }
 
-// class SelectOptionList<T> extends List<SelectOption<T>> {
-//   final List<SelectOption<T>> options;
-
-//   SelectOptionList(this.options);
-
-//   SelectOption getOptionFromValue(T value) {
-//     for (var e in options) {
-//       if (e.value == value) return e;
-//     }
-//     throw Error();
-//   }
-// }
-
-// final SelectOptionList<ThemeMode> selectThemeOptions =
-//     SelectOptionList<ThemeMode>([
-//   SelectOption<ThemeMode>(label: 'ライト', value: ThemeMode.light),
-//   SelectOption<ThemeMode>(label: 'ダーク', value: ThemeMode.dark),
-//   SelectOption<ThemeMode>(label: 'システムのデフォルト', value: ThemeMode.system),
-// ]);
-
-// final SelectOptionList<QrEyeShape> selectQrEyeShapeOptions =
-//     SelectOptionList<QrEyeShape>([
-//   SelectOption<QrEyeShape>(label: 'Circle', value: QrEyeShape.circle),
-//   SelectOption<QrEyeShape>(label: 'Square', value: QrEyeShape.square),
-// ]);
-
-// final SelectOptionList<QrDataModuleShape> selectQrDataModuleShapeOptions =
-//     SelectOptionList<QrDataModuleShape>([
-//   SelectOption<QrDataModuleShape>(
-//       label: 'Circle', value: QrDataModuleShape.circle),
-//   SelectOption<QrDataModuleShape>(
-//       label: 'Square', value: QrDataModuleShape.square),
-// ]);
-
-// final SelectOptionList<int> selectQrErrorCorrectLevelOptions =
-//     SelectOptionList([
-//   SelectOption(label: 'H', value: QrErrorCorrectLevel.H),
-//   SelectOption(label: 'L', value: QrErrorCorrectLevel.L),
-//   SelectOption(label: 'M', value: QrErrorCorrectLevel.M),
-//   SelectOption(label: 'Q', value: QrErrorCorrectLevel.Q),
-// ]);
 final List<SelectOption<ThemeMode>> selectThemeOptions = [
   SelectOption<ThemeMode>(label: 'ライト', value: ThemeMode.light),
   SelectOption<ThemeMode>(label: 'ダーク', value: ThemeMode.dark),
