+++ lib/src/utils.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-// import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 ThemeMode stringToThemeMode(String theme) {
